+++ log/json/logger_test.go
@@ -4,7 +4,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	"os"
+	"io/ioutil"
 
 	"github.com/tidepool-org/platform/log"
 	"github.com/tidepool-org/platform/log/json"
@@ -13,19 +13,19 @@ import (
 var _ = Describe("Logger", func() {
 	Context("NewLogger", func() {
 		It("returns an error if writer is missing", func() {
-			logger, err := json.NewLogger(nil, log.DefaultLevels(), log.DefaultLevel())
+			logger, err := json.NewLogger(nil, log.DefaultLevelRanks(), log.DefaultLevel())
 			Expect(err).To(MatchError("json: writer is missing"))
 			Expect(logger).To(BeNil())
 		})
 
-		It("returns an error if levels is missing", func() {
-			logger, err := json.NewLogger(os.Stdout, nil, log.DefaultLevel())
-			Expect(err).To(MatchError("log: levels is missing"))
+		It("returns an error if level ranks is missing", func() {
+			logger, err := json.NewLogger(ioutil.Discard, nil, log.DefaultLevel())
+			Expect(err).To(MatchError("log: level ranks is missing"))
 			Expect(logger).To(BeNil())
 		})
 
 		It("returns successfully", func() {
-			Expect(json.NewLogger(os.Stdout, log.DefaultLevels(), log.DefaultLevel())).ToNot(BeNil())
+			Expect(json.NewLogger(ioutil.Discard, log.DefaultLevelRanks(), log.DefaultLevel())).ToNot(BeNil())
 		})
 	})
 })